Added missing references to Google Docs, Riak, and PostgreSQL.
Read through introduction and have one comment for Martin: we need examples of the "some" NoSQL databases that adopt a last-writer-wins strategy.

Also: I think we need a "map" of the paper at the end of the introduction, which describes what is contained in each section.  I will add this.